Debug functionallity for LineTracker and AutoAlign
Added JavaDoc comments to the LineTracker methods (/** comments, which show up any time someone hovers over them in cases where they are used -- helpful for someone unfamiliar with the code). Added print statements to LineTracker methods, to assist with debugging. Added a variable to disable debugging print statements to LineTracker. Added a print statement to warn you when the debug print statements were disabled to AutoAlign. Added comments explaining what was going on to AutoAlign. Added print statements to if statements to assist with debugging. Added a pritnln statement to create a new line (the print statements allow for everythign happening in the same interation to be on the same line). Double checked algorithm -- should be ready to go, although the deltaAngle variable needs fixed.